Uses get_reaction_users()
Now gets who reacted with the upvote/downvote when bot was offline. Updates their upvotes/downvotes given and removes offending karma points, updating as needed. However, if people upvote/downvote messages that were sent before the last time Tiger was turned on, when Tiger is turned on those reactions are not accounted for.