Uses timestamps
No longer based off of Tiger's exit message, instead based off the timestamp of the exit command. Commented out the smarts incrementer for each message, so no longer tracks smarts, and commented out smarts command, just in case I want to modify it to be actually useful/funny. I have some janky code at the beginning but it works. Only issue with karma I can think of is if people ask for their karma while the bot is logging on/logging off, at max an 8 second window where the given karma will not be accurate, but I think this is a non-issue.